+++ lib/db/src/schema/index.ts
@@ -13,3 +13,5 @@ export * from "./analyticsEvents";
 export * from "./tokenUsageDaily";
 export * from "./users";
 export * from "./userSessions";
+export * from "./plans";
+export * from "./subscriptions";